rename word-count util to lexical to match usage
