+++ src/app/[orgSlug]/study/[studyId]/edit-and-resubmit/footer.tsx
@@ -6,7 +6,7 @@ import { Button, Group, Stack, Text } from '@mantine/core'
 import { CaretLeftIcon } from '@phosphor-icons/react'
 import { AppModal } from '@/components/modal'
 import { Routes } from '@/lib/routes'
-import { hasLexicalContent } from '@/lib/word-count'
+import { hasLexicalContent } from '@/lib/lexical'
 import { useEditResubmit } from '@/contexts/edit-resubmit'
 import { ReviewerPreview } from '@/app/[orgSlug]/study/[studyId]/proposal/reviewer-preview'
