+++ src/app/[orgSlug]/study/[studyId]/proposal/form.tsx
@@ -12,7 +12,7 @@ import { WordCounter } from '@/components/word-counter'
 import { EditableText } from '@/components/editable-text'
 import ProxyProvider from '@/components/proxy-provider'
 import { DatasetMultiSelect } from '@/components/dataset-multi-select'
-import { countWords } from '@/lib/word-count'
+import { countWords } from '@/lib/lexical'
 import { Routes, ExternalLinks } from '@/lib/routes'
 import { DEFAULT_DRAFT_TITLE, WORD_LIMITS, type ProposalFormValues } from './schema'
 import { useProposal } from '@/contexts/proposal'
