+++ src/app/[orgSlug]/study/[studyId]/proposal/reviewer-preview.tsx
@@ -5,7 +5,7 @@ import { Box, Divider, Stack, Text } from '@mantine/core'
 import { useParams } from 'next/navigation'
 import { EditableText } from '@/components/editable-text'
 import { ResearcherProfilePopover } from '@/components/researcher-profile-popover'
-import { extractTextFromLexical } from '@/lib/word-count'
+import { extractTextFromLexical } from '@/lib/lexical'
 import { useProposal } from '@/contexts/proposal'
 import { useOrgDataSources } from '@/hooks/use-org-data-sources'
 import { usePopover } from '@/hooks/use-popover'
