+++ src/server/actions/study.actions.ts
@@ -5,7 +5,7 @@ import { sql } from 'kysely'
 import { ActionFailure, throwNotFound } from '@/lib/errors'
 import { ActionSuccessType, jobFileSchema } from '@/lib/types'
 import type { StudyStatus } from '@/database/types'
-import { countWordsFromLexical, lexicalJson } from '@/lib/word-count'
+import { countWordsFromLexical, lexicalJson } from '@/lib/lexical'
 import { FEEDBACK_MAX_WORDS, FEEDBACK_MIN_WORDS, toReviewDecision, type Decision } from '@/lib/proposal-review'
 import { reviewFeedbackDocName } from '@/lib/collaboration-documents'
 import { sleep } from '@/lib/utils'
